Added my star-cubing algorithm and data csv file
My algorithm mainly runs based on a csv file. To run star-cubing create a new starCube object with the file name and min_supp value. For example starCube('data.csv', 2) would be an acceptable object.

The created object is in tree form which will need to be traversed in a tree fashion if I understand correctly. Further implementation to traverse tree will most likely be needed.